dev: Rename JS forms controllers TicketEditor and NewAuthorizationForm
The TicketEditorController wasn't a clear name. It is now more explicit
and more precise as it is used only on the "new answer" form (on the
ticket page).

Prefixing the controllers' names by Form allows to group them with the
FormPriorityController.
