+++ assets/javascripts/controllers/form_new_answer_controller.js
