+++ assets/javascripts/controllers/form_new_authorization_controller.js
